pyproject.toml
[project]
name = "agent-bus-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aio-pika>=9.6.2",
    "chromadb>=1.5.5",
    "fastapi>=0.135.2",
    "google-genai>=1.69.0",
    "mcp[cli]>=1.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "honcho>=2.0.0",
    "taskipy>=1.14.1",
]

[tool.taskipy.tasks]
# Sobe o docker em background e depois inicia os agentes com honcho
dev = "docker compose up -d && honcho start"

# Atalho para ver o painel do Rabbit no navegador
rabbit-ui = "python -m webbrowser http://localhost:15672"

# Para tudo, inclusive o docker
stop = "honcho stop; docker compose down"

# Abre o chat para você conversar com a rede
chat = "python interface_agent.py"

# Mantem os processos individuais para debugar algo isolado
bus = "python bus.py"
worker = "python worker_agent.py"
